backend/api-gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
pydantic==2.7.0
pydantic-settings==2.3.0
redis==5.0.4
httpx==0.27.0
slowapi==0.1.9
structlog==24.2.0
